Move context struct to two_phase submodule
Move the two phase commit specific context struct and other associated
structs and implementations to a context module in the two_phase
module.
