+++ services/scabbard/libscabbard/src/store/scabbard_store/diesel/mod.rs
@@ -699,10 +699,10 @@ pub mod tests {
 
     use crate::store::scabbard_store::{
         commit::{CommitEntryBuilder, ConsensusDecision},
-        context::{ContextBuilder, Participant},
         service::{ConsensusType, ScabbardServiceBuilder, ServiceStatus},
         two_phase::{
             action::{ConsensusAction, ConsensusActionNotification},
+            context::{ContextBuilder, Participant},
             event::Scabbard2pcEvent,
             message::Scabbard2pcMessage,
             state::Scabbard2pcState,
